requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pyotp==2.9.0
PyYAML==6.0.2
